+++ packages/nx-plugin/src/generators/init/lib/add-analog-dependencies.ts
@@ -6,10 +6,21 @@ import { getAnalogDependencies } from '../../../utils/versions/dependencies';
 export function addAnalogDependencies(
   tree: Tree,
   angularVersion: string,
+  vitest: boolean,
   nxVersion?: string,
 ) {
   const devDependencies = getAnalogDevDependencies(angularVersion, nxVersion);
   const dependencies = getAnalogDependencies(angularVersion);
+
+  if (!nxVersion) {
+    delete devDependencies['@nx/vite'];
+  }
+
+  if (!vitest) {
+    delete devDependencies['vitest'];
+    delete devDependencies['@vitest/coverage-v8'];
+    delete devDependencies['@vitest/ui'];
+  }
 
   addDependenciesToPackageJson(
     tree,