+++ packages/nx-plugin/src/generators/preset/generator.spec.ts
@@ -48,4 +48,20 @@ describe('preset generator', () => {
 
     expect(tree.read('/my-app/src/test-setup.ts').toString()).toMatchSnapshot();
   });
+
+  it('should include package.json deps', async () => {
+    const { tree } = await setup({ analogAppName: 'my-app' });
+
+    const packageJson = JSON.parse(tree.read('/package.json').toString());
+
+    expect(packageJson['devDependencies']['@nx/vite']).toBeDefined();
+  });
+
+  it('should use vitest 3 for Nx < 22.3.0', async () => {
+    const { tree } = await setup({ analogAppName: 'my-app' });
+
+    const packageJson = JSON.parse(tree.read('/package.json').toString());
+
+    expect(packageJson['devDependencies']['vitest']).toBe('^3.0.0');
+  });
 });