+++ packages/nx-plugin/src/generators/preset/generator.ts
@@ -3,6 +3,7 @@ import { PresetGeneratorSchema } from './schema';
 
 export default async function (tree: Tree, options: PresetGeneratorSchema) {
   ensurePackage('@nx/angular', NX_VERSION);
+  ensurePackage('@nx/vite', NX_VERSION);
   ensurePackage('@angular-devkit/core', 'latest');
   ensurePackage('rxjs', 'latest');