+++ packages/nx-plugin/src/utils/versions/dev-dependencies.ts
@@ -104,6 +104,7 @@ const getDevDependencies = (
     '@analogjs/platform': V19_X_ANALOG_JS_PLATFORM,
     '@analogjs/vite-plugin-angular': V19_X_ANALOG_JS_VITE_PLUGIN_ANGULAR,
     '@analogjs/vitest-angular': V19_X_ANALOG_JS_VITEST_ANGULAR,
+    '@nx/vite': V19_X_NX_VITE,
     jsdom: V19_X_JSDOM,
     'vite-tsconfig-paths': V19_X_VITE_TSCONFIG_PATHS,
     vite:
@@ -111,15 +112,18 @@ const getDevDependencies = (
         ? V19_X_VITE
         : NX_X_LATEST_VITE,
     vitest:
-      escapedNxVersion && lt(escapedNxVersion, '22.3.0')
+      lt(escapedAngularVersion, '21.0.0') ||
+      (escapedNxVersion && lt(escapedNxVersion, '22.3.0'))
         ? V19_X_VITEST
         : NX_X_LATEST_VITEST,
     '@vitest/coverage-v8':
-      escapedNxVersion && lt(escapedNxVersion, '22.3.0')
+      lt(escapedAngularVersion, '21.0.0') ||
+      (escapedNxVersion && lt(escapedNxVersion, '22.3.0'))
         ? V19_X_VITEST
         : NX_X_LATEST_VITEST,
     '@vitest/ui':
-      escapedNxVersion && lt(escapedNxVersion, '22.3.0')
+      lt(escapedAngularVersion, '21.0.0') ||
+      (escapedNxVersion && lt(escapedNxVersion, '22.3.0'))
         ? V19_X_VITEST
         : NX_X_LATEST_VITEST,
   };